+++ src/unoserver/comparer.py
@@ -13,6 +13,7 @@ import logging
 import os
 import sys
 import unohelper
+import warnings
 
 from com.sun.star.beans import PropertyValue
 from com.sun.star.io import XOutputStream
@@ -277,13 +278,46 @@ def main():
     )
     parser.add_argument(
         "--convert-to",
+        help="The file type/extension of the output file (ex pdf). Deprecated in favor for --file-type",
+    )
+    parser.add_argument(
+        "--file-type",
         help="The file type/extension of the output file (ex pdf). Required when using stdout",
     )
     parser.add_argument(
-        "--interface", default="127.0.0.1", help="The interface used by the server"
+        "--interface",
+        default=None,
+        help="The interface used by the server. Deprecated in favor for --host",
+    )
+    parser.add_argument(
+        "--host", default="127.0.0.1", help="The host used by the server"
     )
     parser.add_argument("--port", default="2002", help="The port used by the server")
     args = parser.parse_args()
+
+    if not sys.warnoptions:
+        warnings.simplefilter("default")  # Change the filter in this process
+
+    warnings.warn(
+        "Note that the order of the file parameters will change in 2.0.",
+        DeprecationWarning,
+    )
+
+    if args.interface is not None:
+        warnings.warn(
+            "The argument --interface has been renamed --host and will stop working in 2.0.",
+            DeprecationWarning,
+        )
+    if args.interface is None and args.host is not None:
+        args.interface = args.host
+
+    if args.convert_to is not None:
+        warnings.warn(
+            "The argument --convert-to has been renamed --file-type and will stop working in 2.0.",
+            DeprecationWarning,
+        )
+    if args.file_type is not None:
+        args.convert_to = args.file_type
 
     comparer = UnoComparer(args.interface, args.port)